BACKWARD-INCOMPATIBLE: Moved from custom save method to auto_now(_add).
This may be an arbitrary change, but I want less code that to be
accountable for. Using the auto_now attributes means fewer tests
to write.

The main difference is that now Topic and Question `modified`
fields will never be `NULL`. Rather, they will always have a date.
Previously, an item that was created but never modified did not
have a modified date.