Change to the quasi-standard __version__ for providing package version.
Also use a reliable way to get version into setup.py. Depending on what
happens in the package __init__.py, you can have import issues when
importing into the setup.py, when setup is not complete. This way is
slightly more complex, but avoids the import issues. Thanks to carljm
for the example.